+++ recipes/docker.rb
@@ -32,6 +32,11 @@ if is_node_server || node['cookbook-openshift3']['deploy_containerized']
     source 'docker-storage.erb'
   end
 
+  template '/etc/sysconfig/docker-network' do
+    source 'service_docker-network.sysconfig.erb'
+    notifies :restart, 'service[docker]', :immediately unless ::Mixlib::ShellOut.new('systemctl is-enabled docker').run_command.error?
+  end
+
   template '/etc/sysconfig/docker' do
     source 'service_docker.sysconfig.erb'
     notifies :restart, 'service[docker]', :immediately